Replaced switch statement of hard-coded opcodes with logic that utilizes opcode patterns.
For example, instead of looping over codes and checking each against a long list of possible opcodes:

    switch (code) {
        case LD_A:
            ...
        case LD_B:
            ...
    ...

the opcode is now decoded into values x, y, z, p, and q and matched against the tables found here (http://www.z80.info/decoding.htm) to determine the appropriate instruction. An array of pointers to registers is used to create the "r" table in the given document.

Just enough is implemented to load 8-bit registers and to HALT, with all other opcodes passed over and announced as unknown.